+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## v1.0
+### v1.0.1
+- Added 12 icons, updated 4 icons.
+- Added convenience method `setSelectedIcon(int iconId)`` for selecting a single icon
+- Change `setAllowMultipleSelection` to `setMaxSelection`
+- Removed BroadcastReceiver for language change, it was not unregistered
+- Search bar now has search action and keyboard is hidden on enter.
+- Synchronized some methods
+
+## v1.0.0
 - Initial release
 - Customizable styles for almost every part of the dialog
 - Many available options